pyproject.toml
[tool.poetry]
name = "orchestrator"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
mss = "^10.0.0"
pillow = "^11.1.0"
g4f = "^0.4.7.6"
docker = "^7.1.0"
sqlalchemy = "^2.0.38"
flask = "^3.1.0"
flask-cors = "^5.0.1"
click = "^8.1.8"
apscheduler = "^3.11.0"
gradio = "^5.21.0"
openai = "^1.66.3"
anthropic = "^0.49.0"
dotenv = "^0.9.9"
asyncvnc = "^1.3.0"

[tool.poetry.scripts]
start = "orchestrator.test:main"
gui = "orchestrator.gui:main"
flask = "orchestrator.app:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
